Added folder/file input management
Added the possibility to add file and folder paths to tasks and subtasks. The path will be displayed in a reduced version under the input field, indicating with a logo their folder or file, and are clickable.
When clicking to add a file or folder, a deviec-specific dialog window will open choose the path.
There seem to be a small visual problem with the add file/folder part of the form on a small android phone : visual overlapping.