Added periodicity management
Added a basic periodicity input manager, where the user can input, in days, a periodicity for a cycling task.
After looking into the Duration Class, it seems it can only handle periods as a value in millisecond, and therefore may not be able to take into account a "every month" or "every year" concepts, as length of month and years may vary.
It is also questionnable as to how the periodicity will be handled in the views and point management.