Added a test notification system
Added a basic notification system, currently setup to test its usage and correct response. This system needs to be clean and added to the basic notification configuration setup for it to be working properly. As of now, instant notifications as well as timed notifications have been setup. However, it is not entirely sure if the notifications work in all environment, and do not return errors in non working environments. To cover for android 13+ new configuration, special request for exact alarm are made, and inexact ones are use if exact is impossible. Also, it seems that executed notifications stay in the pending notification list, but it unsure if it is a normal interaction.