Created a "detailled view" page, and fixed bugs
Created a detailled view page, which gives all the necessary and additional information of a task, and give the option to mark as completed a task or subtask.
Added a button within task card to access the detailled view of said task.
Fixed display issue of general task list view page with small screen, and problem with double scrollable area.
Fixed database issue with id management and subtask extraction.